+++ R/render.r
@@ -76,7 +76,8 @@ render_page <- function(pkg = ".", name, data, path = "", depth = NULL, quiet = 
 data_template <- function(pkg = ".", depth = 0L) {
   pkg <- as_pkgdown(pkg)
 
-  authors <- data_authors(pkg)$main %>%
+  roles <- pkg$meta$authors$footer$roles %||% default_roles()
+  authors <- data_authors(pkg, roles = roles)$main %>%
     purrr::map_chr("name") %>%
     paste(collapse = ", ")
 
@@ -311,7 +312,7 @@ check_made_by <- function(first) {
 pkgdown_footer <- function(data, pkg) {
 
   footer_components <- list(
-    authors = footer_authors(data),
+    authors = footer_authors(data, pkg),
     pkgdown = footer_pkgdown(data)
   )
 
@@ -358,8 +359,9 @@ pkgdown_footer <- function(data, pkg) {
   list(left = left_final_components, right = right_final_components)
 }
 
-footer_authors <- function(data) {
-  paste0("Developed by ", data$package$authors, ".")
+footer_authors <- function(data, pkg) {
+  text <- pkg$meta$authors$footer$text %||% "Developed by"
+  paste0(trimws(text), " ", data$package$authors, ".")
 }
 
 footer_pkgdown <- function(data) {