+++ tests/testthat/test-build-home-authors.R
@@ -33,3 +33,27 @@ test_that("Comments in authors info are linkified", {
     "&lt;<a href='https://x.org/'>https://x.org/</a>&gt;)</small>"
   )
 })
+
+test_that("Data authors can accept different filtering", {
+  pkg <- as_pkgdown(test_path("assets/sidebar"))
+  expect_length(data_authors(pkg)$main, 2)
+  expect_length(data_authors(pkg, roles = "cre")$main, 1)
+})
+
+test_that("Text can be added", {
+  pkg <- as_pkgdown(test_path("assets/sidebar-comment"))
+  expect_null(data_authors_page(pkg)$after)
+  expect_null(data_authors_page(pkg)$before)
+
+  pkg$meta$authors$before <- "Dream team:"
+  pkg$meta$authors$after <- "You are welcome!"
+  expect_equal(data_authors_page(pkg)$before, "Dream team:")
+  expect_equal(data_authors_page(pkg)$after, "You are welcome!")
+})
+
+test_that("data_home_sidebar_authors() works with text", {
+  pkg <- as_pkgdown(test_path("assets/sidebar-comment"))
+  pkg$meta$authors$sidebar$before <- "yay"
+  pkg$meta$authors$sidebar$after <- "cool"
+  expect_snapshot(cat(data_home_sidebar_authors(pkg)))
+})