+++ tests/testthat/test-data_home_sidebar.R
@@ -2,6 +2,19 @@ test_that("data_home_sidebar() works by default", {
   pkg <- test_path("assets/sidebar")
   pkg <- as_pkgdown(pkg)
   expect_snapshot(cat(data_home_sidebar(pkg)))
+
+  pkg <- test_path("assets/sidebar-comment")
+  pkg <- as_pkgdown(pkg)
+  expect_snapshot(
+    cat(
+      as.character(
+        xml2::xml_find_first(
+          xml2::read_html(data_home_sidebar(pkg)),
+          ".//div[@class='developers']"
+        )
+      )
+    )
+  )
 })
 
 test_that("data_home_sidebar() can be removed", {