+++ NEWS.md
@@ -1,5 +1,14 @@
 # pkgdown (development version)
 
+* Authors configuration is more flexible. Users can now:
+  * Choose the roles used for filtering authors for the sidebar and footer.
+  * Choose the text before authors in the footer.
+  * Add text before and after the authors list in the sidebar.
+  * Add text before and after the authors list of the authors page. (#1516)
+  
+* Make authors' non-ORCID comments from DESCRIPTION more usable as bio/description of 
+  contributions: add a link to the authors page from the sidebar if any author
+  has a non-ORCID comment, and only render non-ORCID comments on the authors page. (#1516)
 * Better handling for mix of citations with and without text version. Also
   escapes HTML in the text version (@bastistician, #1507).